Extracción de las funciones de App.js
Se ha sacado la definición de las funciones usadas en App.js para una mejor legibilidad del código
