+++ mybuttonproject.web/src/lists_config.json
@@ -1,24 +1,19 @@
 {
     "listOrder" : ["chistes", "quijote"],
-
     "chistes" : {
-
         "nombreFichero" : "chistes",
         "titulo" : "Bienvenido al botón de chistes aleatorios",
         "textoBoton" : ["Cuéntame un chiste", "Respuesta", "Otro chiste"],
         "tipoLista" : "ListChistes",
-        "saltoPagina" : 0
-        
-
+        "espera" : 1
     },
     "quijote" : {
         "nombreFichero" : "chistes",
         "titulo" : "Suficientes risas, pongámonos serios",
         "textoBoton" : ["Leer el Quijote", "Siguiente"],
         "tipoLista" : "ListQuijote",
-        "saltoPagina" : 1
+        "espera" : 0
     }
-
 }
